+++ x/cronos/types/messages.go
@@ -12,15 +12,7 @@ import (
 	"github.com/ethereum/go-ethereum/common"
 )
 
-const (
-	TypeMsgConvertVouchers    = "ConvertVouchers"
-	TypeMsgTransferTokens     = "TransferTokens"
-	TypeMsgUpdateTokenMapping = "UpdateTokenMapping"
-	TypeMsgUpdateParams       = "UpdateParams"
-	TypeMsgTurnBridge         = "TurnBridge"
-	TypeMsgUpdatePermissions  = "UpdatePermissions"
-	TypeMsgStoreBlockList     = "StoreBlockList"
-)
+const TypeMsgUpdateTokenMapping = "UpdateTokenMapping"
 
 var (
 	_ sdk.Msg = &MsgConvertVouchers{}
@@ -37,31 +29,6 @@ func NewMsgConvertVouchers(address string, coins sdk.Coins) *MsgConvertVouchers 
 		Address: address,
 		Coins:   coins,
 	}
-}
-
-// Route ...
-func (msg MsgConvertVouchers) Route() string {
-	return RouterKey
-}
-
-// Type ...
-func (msg MsgConvertVouchers) Type() string {
-	return TypeMsgConvertVouchers
-}
-
-// GetSigners ...
-func (msg *MsgConvertVouchers) GetSigners() []sdk.AccAddress {
-	address, err := sdk.AccAddressFromBech32(msg.Address)
-	if err != nil {
-		panic(err)
-	}
-	return []sdk.AccAddress{address}
-}
-
-// GetSignBytes ...
-func (msg *MsgConvertVouchers) GetSignBytes() []byte {
-	bz := ModuleCdc.MustMarshalJSON(msg)
-	return sdk.MustSortJSON(bz)
 }
 
 // ValidateBasic ...
@@ -88,31 +55,6 @@ func NewMsgTransferTokens(from string, to string, coins sdk.Coins) *MsgTransferT
 		To:    to,
 		Coins: coins,
 	}
-}
-
-// Route ...
-func (msg MsgTransferTokens) Route() string {
-	return RouterKey
-}
-
-// Type ...
-func (msg MsgTransferTokens) Type() string {
-	return TypeMsgTransferTokens
-}
-
-// GetSigners ...
-func (msg *MsgTransferTokens) GetSigners() []sdk.AccAddress {
-	from, err := sdk.AccAddressFromBech32(msg.From)
-	if err != nil {
-		panic(err)
-	}
-	return []sdk.AccAddress{from}
-}
-
-// GetSignBytes ...
-func (msg *MsgTransferTokens) GetSignBytes() []byte {
-	bz := ModuleCdc.MustMarshalJSON(msg)
-	return sdk.MustSortJSON(bz)
 }
 
 // ValidateBasic ...
@@ -174,20 +116,9 @@ func (msg *MsgUpdateTokenMapping) ValidateBasic() error {
 	return nil
 }
 
-// Route ...
-func (msg MsgUpdateTokenMapping) Route() string {
-	return RouterKey
-}
-
 // Type ...
 func (msg MsgUpdateTokenMapping) Type() string {
 	return TypeMsgUpdateTokenMapping
-}
-
-// GetSignBytes ...
-func (msg *MsgUpdateTokenMapping) GetSignBytes() []byte {
-	bz := ModuleCdc.MustMarshalJSON(msg)
-	return sdk.MustSortJSON(bz)
 }
 
 // NewMsgTurnBridge ...
@@ -196,15 +127,6 @@ func NewMsgTurnBridge(admin string, enable bool) *MsgTurnBridge {
 		Sender: admin,
 		Enable: enable,
 	}
-}
-
-// GetSigners ...
-func (msg *MsgTurnBridge) GetSigners() []sdk.AccAddress {
-	sender, err := sdk.AccAddressFromBech32(msg.Sender)
-	if err != nil {
-		panic(err)
-	}
-	return []sdk.AccAddress{sender}
 }
 
 // ValidateBasic ...
@@ -217,36 +139,11 @@ func (msg *MsgTurnBridge) ValidateBasic() error {
 	return nil
 }
 
-// Route ...
-func (msg MsgTurnBridge) Route() string {
-	return RouterKey
-}
-
-// Type ...
-func (msg MsgTurnBridge) Type() string {
-	return TypeMsgTurnBridge
-}
-
-// GetSignBytes ...
-func (msg *MsgTurnBridge) GetSignBytes() []byte {
-	bz := ModuleCdc.MustMarshalJSON(msg)
-	return sdk.MustSortJSON(bz)
-}
-
 func NewMsgUpdateParams(authority string, params Params) *MsgUpdateParams {
 	return &MsgUpdateParams{
 		Authority: authority,
 		Params:    params,
 	}
-}
-
-// GetSigners returns the expected signers for a MsgUpdateParams message.
-func (msg *MsgUpdateParams) GetSigners() []sdk.AccAddress {
-	addr, err := sdk.AccAddressFromBech32(msg.Authority)
-	if err != nil {
-		panic(err)
-	}
-	return []sdk.AccAddress{addr}
 }
 
 // ValidateBasic does a sanity check on the provided data.
@@ -262,22 +159,6 @@ func (msg *MsgUpdateParams) ValidateBasic() error {
 	return nil
 }
 
-// Route ...
-func (msg MsgUpdateParams) Route() string {
-	return RouterKey
-}
-
-// Type ...
-func (msg MsgUpdateParams) Type() string {
-	return TypeMsgUpdateParams
-}
-
-// GetSignBytes ...
-func (msg *MsgUpdateParams) GetSignBytes() []byte {
-	bz := ModuleCdc.MustMarshalJSON(msg)
-	return sdk.MustSortJSON(bz)
-}
-
 // NewMsgUpdatePermissions ...
 func NewMsgUpdatePermissions(from string, address string, permissions uint64) *MsgUpdatePermissions {
 	return &MsgUpdatePermissions{
@@ -285,15 +166,6 @@ func NewMsgUpdatePermissions(from string, address string, permissions uint64) *M
 		Address:     address,
 		Permissions: permissions,
 	}
-}
-
-// GetSigners ...
-func (msg *MsgUpdatePermissions) GetSigners() []sdk.AccAddress {
-	sender, err := sdk.AccAddressFromBech32(msg.From)
-	if err != nil {
-		panic(err)
-	}
-	return []sdk.AccAddress{sender}
 }
 
 // ValidateBasic ...
@@ -308,22 +180,6 @@ func (msg *MsgUpdatePermissions) ValidateBasic() error {
 	}
 
 	return nil
-}
-
-// Route ...
-func (msg MsgUpdatePermissions) Route() string {
-	return RouterKey
-}
-
-// Type ...
-func (msg MsgUpdatePermissions) Type() string {
-	return TypeMsgUpdatePermissions
-}
-
-// GetSignBytes ...
-func (msg *MsgUpdatePermissions) GetSignBytes() []byte {
-	bz := ModuleCdc.MustMarshalJSON(msg)
-	return sdk.MustSortJSON(bz)
 }
 
 func NewMsgStoreBlockList(from string, blob []byte) *MsgStoreBlockList {
@@ -346,35 +202,11 @@ func (msg *MsgStoreBlockList) ValidateBasic() error {
 	if err != nil {
 		return errors.Wrapf(sdkerrors.ErrInvalidAddress, "invalid sender address (%s)", err)
 	}
-
+	// skip heavy operation in Decrypt by early return with errDummyIdentity in
+	// https://github.com/FiloSottile/age/blob/v1.1.1/age.go#L197
 	_, err = age.Decrypt(bytes.NewBuffer(msg.Blob), new(dummyIdentity))
 	if err != nil && err != errDummyIdentity {
 		return err
 	}
 	return nil
-}
-
-func (msg *MsgStoreBlockList) GetSigners() []sdk.AccAddress {
-	addr, err := sdk.AccAddressFromBech32(msg.From)
-	if err != nil {
-		panic(err)
-	}
-
-	return []sdk.AccAddress{addr}
-}
-
-// GetSignBytes ...
-func (msg *MsgStoreBlockList) GetSignBytes() []byte {
-	bz := ModuleCdc.MustMarshalJSON(msg)
-	return sdk.MustSortJSON(bz)
-}
-
-// Route ...
-func (msg MsgStoreBlockList) Route() string {
-	return RouterKey
-}
-
-// Type ...
-func (msg MsgStoreBlockList) Type() string {
-	return TypeMsgStoreBlockList
 }