+++ x/e2ee/types/msg.go
@@ -21,11 +21,3 @@ func ValidateRecipientKey(key string) error {
 	_, err := age.ParseX25519Recipient(key)
 	return err
 }
-
-func (m *MsgRegisterEncryptionKey) GetSigners() []sdk.AccAddress {
-	addr, err := sdk.AccAddressFromBech32(m.Address)
-	if err != nil {
-		panic(err)
-	}
-	return []sdk.AccAddress{addr}
-}